requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
gunicorn==23.0.0
ultralytics==8.0.185
pillow==10.1.0
torchvision==0.17.1
numpy==1.26.0
